chore: clean up search tool renderer and remnants of simple agent framework
