+++ web/src/app/chat/message/messageComponents/renderers/SearchToolRenderer.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState, useRef, useMemo } from "react";
-import { FiSearch, FiGlobe } from "react-icons/fi";
+import React, { useEffect, useState, useRef, useMemo, JSX } from "react";
+import { FiSearch, FiGlobe, FiBookOpen } from "react-icons/fi";
 import {
   PacketType,
   SearchToolPacket,
@@ -8,35 +8,36 @@ import {
   SearchToolDocumentsDelta,
   SectionEnd,
 } from "../../../services/streamingModels";
-import { MessageRenderer } from "../interfaces";
-import { ResultIcon } from "@/components/chat/sources/SourceCard";
-import { OnyxDocument } from "@/lib/search/interfaces";
+import { MessageRenderer, RendererResult } from "../interfaces";
+import { truncateString } from "@/lib/utils";
 import { SourceChip2 } from "@/app/chat/components/SourceChip2";
 import { BlinkingDot } from "../../BlinkingDot";
-import Text from "@/refresh-components/texts/Text";
-import { SearchToolRendererV2 } from "./SearchToolRendererV2";
-import { usePostHog } from "posthog-js/react";
-import { ResearchType } from "@/app/chat/interfaces";
-import { clearTimeoutRefs } from "../timing";
+import { OnyxDocument } from "@/lib/search/interfaces";
+import { ResultIcon } from "@/components/chat/sources/SourceCard";
 
-const INITIAL_RESULTS_TO_SHOW = 3;
-const RESULTS_PER_EXPANSION = 10;
+const MAX_TITLE_LENGTH = 25;
 
 const INITIAL_QUERIES_TO_SHOW = 3;
 const QUERIES_PER_EXPANSION = 5;
 
+const INITIAL_RESULTS_TO_SHOW = 3;
+const RESULTS_PER_EXPANSION = 10;
+
 const SEARCHING_MIN_DURATION_MS = 1000; // 1 second minimum for "Searching" state
 const SEARCHED_MIN_DURATION_MS = 1000; // 1 second minimum for "Searched" state
 
-const constructCurrentSearchState = (
-  packets: SearchToolPacket[]
-): {
+export interface SearchState {
   queries: string[];
   results: OnyxDocument[];
   isSearching: boolean;
+  hasResults: boolean;
   isComplete: boolean;
   isInternetSearch: boolean;
-} => {
+}
+
+export const constructCurrentSearchState = (
+  packets: SearchToolPacket[]
+): SearchState => {
   // Check for new specific search tool packets first
   const searchStart = packets.find(
     (packet) => packet.obj.type === PacketType.SEARCH_TOOL_START
@@ -65,6 +66,7 @@ const constructCurrentSearchState = (
     .flatMap((delta) => delta?.queries || [])
     .filter((query, index, arr) => arr.indexOf(query) === index); // Remove duplicates
 
+  // Extract documents/results from document delta packets
   const seenDocIds = new Set<string>();
   const results = documentDeltas
     .flatMap((delta) => delta?.documents || [])
@@ -76,28 +78,161 @@ const constructCurrentSearchState = (
     });
 
   const isSearching = Boolean(searchStart && !searchEnd);
+  const hasResults = results.length > 0;
   const isComplete = Boolean(searchStart && searchEnd);
   const isInternetSearch = searchStart?.is_internet_search || false;
 
-  return { queries, results, isSearching, isComplete, isInternetSearch };
+  return {
+    queries,
+    results,
+    isSearching,
+    hasResults,
+    isComplete,
+    isInternetSearch,
+  };
 };
 
-export const SearchToolRenderer: MessageRenderer<
-  SearchToolPacket,
-  { researchType?: string | null }
-> = ({
+// Step 1 Renderer: "Searching internally" with queries
+export function SearchToolStep1Renderer({
   packets,
-  state,
+  isActive,
+  children,
+}: {
+  packets: SearchToolPacket[];
+  isActive: boolean;
+  children: (result: RendererResult) => JSX.Element;
+}) {
+  const { queries } = constructCurrentSearchState(packets);
+  const [queriesToShow, setQueriesToShow] = useState(INITIAL_QUERIES_TO_SHOW);
+
+  return children({
+    icon: FiSearch,
+    status: "Searching internally",
+    content: (
+      <div className="flex flex-col">
+        <div className="flex flex-wrap gap-x-2 gap-y-2 ml-1">
+          {queries.slice(0, queriesToShow).map((query, index) => (
+            <div
+              key={index}
+              className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+              style={{
+                animationDelay: `${index * 30}ms`,
+                animationFillMode: "backwards",
+              }}
+            >
+              <SourceChip2
+                icon={<FiSearch size={10} />}
+                title={truncateString(query, MAX_TITLE_LENGTH)}
+              />
+            </div>
+          ))}
+          {queries.length > queriesToShow && (
+            <div
+              className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+              style={{
+                animationDelay: `${queriesToShow * 30}ms`,
+                animationFillMode: "backwards",
+              }}
+            >
+              <SourceChip2
+                title={`${queries.length - queriesToShow} more...`}
+                onClick={() => {
+                  setQueriesToShow((prevQueries) =>
+                    Math.min(
+                      prevQueries + QUERIES_PER_EXPANSION,
+                      queries.length
+                    )
+                  );
+                }}
+              />
+            </div>
+          )}
+          {queries.length === 0 && <BlinkingDot />}
+        </div>
+      </div>
+    ),
+  });
+}
+
+// Step 2 Renderer: "Reading" with documents
+export function SearchToolStep2Renderer({
+  packets,
+  isActive,
+  children,
+}: {
+  packets: SearchToolPacket[];
+  isActive: boolean;
+  children: (result: RendererResult) => JSX.Element;
+}) {
+  const { results } = constructCurrentSearchState(packets);
+  const [resultsToShow, setResultsToShow] = useState(INITIAL_RESULTS_TO_SHOW);
+
+  return children({
+    icon: FiBookOpen,
+    status: "Reading",
+    content: (
+      <div className="flex flex-col">
+        <div className="flex flex-wrap gap-x-2 gap-y-2 ml-1">
+          {results.slice(0, resultsToShow).map((result, index) => (
+            <div
+              key={result.document_id}
+              className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+              style={{
+                animationDelay: `${index * 30}ms`,
+                animationFillMode: "backwards",
+              }}
+            >
+              <SourceChip2
+                icon={<ResultIcon doc={result} size={10} />}
+                title={truncateString(
+                  result.semantic_identifier || "",
+                  MAX_TITLE_LENGTH
+                )}
+                onClick={() => {
+                  if (result.link) {
+                    window.open(result.link, "_blank");
+                  }
+                }}
+              />
+            </div>
+          ))}
+          {results.length > resultsToShow && (
+            <div
+              className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+              style={{
+                animationDelay: `${
+                  Math.min(resultsToShow, results.length) * 30
+                }ms`,
+                animationFillMode: "backwards",
+              }}
+            >
+              <SourceChip2
+                title={`${results.length - resultsToShow} more...`}
+                onClick={() => {
+                  setResultsToShow((prevResults) =>
+                    Math.min(
+                      prevResults + RESULTS_PER_EXPANSION,
+                      results.length
+                    )
+                  );
+                }}
+              />
+            </div>
+          )}
+          {results.length === 0 && <BlinkingDot />}
+        </div>
+      </div>
+    ),
+  });
+}
+
+export const SearchToolRenderer: MessageRenderer<SearchToolPacket, {}> = ({
+  packets,
   onComplete,
-  renderType,
   animate,
   stopPacketSeen,
   children,
 }) => {
-  // Check if this message has a research_type, which indicates it's using the simple agent framework
-  const isDeepResearch = state.researchType === ResearchType.Deep;
-
-  // Initialize all hooks at the top level (before any conditional returns)
   const { queries, results, isSearching, isComplete, isInternetSearch } =
     constructCurrentSearchState(packets);
 
@@ -110,11 +245,11 @@ export const SearchToolRenderer: MessageRenderer<
   const searchedTimeoutRef = useRef<NodeJS.Timeout | null>(null);
   const completionHandledRef = useRef(false);
 
-  // Track how many results to show
-  const [resultsToShow, setResultsToShow] = useState(INITIAL_RESULTS_TO_SHOW);
-
   // Track how many queries to show
   const [queriesToShow, setQueriesToShow] = useState(INITIAL_QUERIES_TO_SHOW);
+
+  // Track how many results to show
+  const [resultsToShow, setResultsToShow] = useState(INITIAL_RESULTS_TO_SHOW);
 
   // Track when search starts (even if the search completes instantly)
   useEffect(() => {
@@ -136,7 +271,12 @@ export const SearchToolRenderer: MessageRenderer<
       // If stopped, skip intermediate states and complete immediately
       if (stopPacketSeen) {
         // Clear any pending timeouts
-        clearTimeoutRefs([timeoutRef, searchedTimeoutRef]);
+        if (timeoutRef.current) {
+          clearTimeout(timeoutRef.current);
+        }
+        if (searchedTimeoutRef.current) {
+          clearTimeout(searchedTimeoutRef.current);
+        }
 
         // Skip "Searched" state, go directly to completion
         setShouldShowAsSearching(false);
@@ -183,7 +323,14 @@ export const SearchToolRenderer: MessageRenderer<
   // Cleanup timeouts when stopped
   useEffect(() => {
     if (stopPacketSeen) {
-      clearTimeoutRefs([timeoutRef, searchedTimeoutRef], true);
+      if (timeoutRef.current) {
+        clearTimeout(timeoutRef.current);
+        timeoutRef.current = null;
+      }
+      if (searchedTimeoutRef.current) {
+        clearTimeout(searchedTimeoutRef.current);
+        searchedTimeoutRef.current = null;
+      }
       // Reset states to prevent flickering
       setShouldShowAsSearching(false);
       setShouldShowAsSearched(false);
@@ -193,7 +340,12 @@ export const SearchToolRenderer: MessageRenderer<
   // Cleanup timeouts on unmount
   useEffect(() => {
     return () => {
-      clearTimeoutRefs([timeoutRef, searchedTimeoutRef]);
+      if (timeoutRef.current) {
+        clearTimeout(timeoutRef.current);
+      }
+      if (searchedTimeoutRef.current) {
+        clearTimeout(searchedTimeoutRef.current);
+      }
     };
   }, []);
 
@@ -223,22 +375,6 @@ export const SearchToolRenderer: MessageRenderer<
   // Determine the icon based on search type
   const icon = isInternetSearch ? FiGlobe : FiSearch;
 
-  // Use V2 renderer unless feature flag is enabled to disable it, and not deep research
-  if (!isDeepResearch) {
-    return (
-      <SearchToolRendererV2
-        packets={packets}
-        state={state}
-        onComplete={onComplete}
-        renderType={renderType}
-        animate={animate}
-        stopPacketSeen={stopPacketSeen}
-      >
-        {children}
-      </SearchToolRendererV2>
-    );
-  }
-
   // Don't render anything if search hasn't started
   if (queries.length === 0) {
     return children({
@@ -252,107 +388,109 @@ export const SearchToolRenderer: MessageRenderer<
     icon,
     status,
     content: (
-      <div className="flex flex-col py-3 gap-2">
-        <Text text02 secondaryBody>
-          Queries
-        </Text>
-        <div className="flex flex-wrap gap-2 pl-1">
-          {queries.slice(0, queriesToShow).map((query, index) => (
-            <div
-              key={index}
-              className="animate-in fade-in slide-in-from-left-2 duration-150"
-              style={{
-                animationDelay: `${index * 30}ms`,
-                animationFillMode: "backwards",
-              }}
-            >
-              <SourceChip2 icon={<FiSearch size={10} />} title={query} />
-            </div>
-          ))}
-          {/* Show a blurb if there are more queries than we are displaying */}
-          {queries.length > queriesToShow && (
-            <div
-              className="animate-in fade-in slide-in-from-left-2 duration-150"
-              style={{
-                animationDelay: `${queriesToShow * 30}ms`,
-                animationFillMode: "backwards",
-              }}
-            >
-              <SourceChip2
-                title={`${queries.length - queriesToShow} more...`}
-                onClick={() => {
-                  setQueriesToShow((prevQueries) =>
-                    Math.min(
-                      prevQueries + QUERIES_PER_EXPANSION,
-                      queries.length
-                    )
-                  );
+      <div className="flex flex-col mt-1.5">
+        <div className="flex flex-col">
+          <div className="flex flex-wrap gap-x-2 gap-y-2 ml-1">
+            {queries.slice(0, queriesToShow).map((query, index) => (
+              <div
+                key={index}
+                className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+                style={{
+                  animationDelay: `${index * 30}ms`,
+                  animationFillMode: "backwards",
                 }}
-              />
-            </div>
-          )}
-        </div>
-
-        {/* If no queries, show a loading state */}
-        {queries.length === 0 && <BlinkingDot />}
-
-        <Text text02 secondaryBody>
-          {isInternetSearch ? "Results" : "Documents"}
-        </Text>
-
-        <div className="flex flex-wrap gap-2 ml-1">
-          {results.slice(0, resultsToShow).map((result, index) => (
-            <div
-              key={result.document_id}
-              className="animate-in fade-in slide-in-from-left-2 duration-150"
-              style={{
-                animationDelay: `${index * 30}ms`,
-                animationFillMode: "backwards",
-              }}
-            >
-              <div className="text-xs">
+              >
                 <SourceChip2
-                  icon={<ResultIcon doc={result} size={10} />}
-                  title={result.semantic_identifier || ""}
-                  onClick={() => {
-                    if (result.link) {
-                      window.open(result.link, "_blank");
-                    }
-                  }}
+                  icon={<FiSearch size={10} />}
+                  title={truncateString(query, MAX_TITLE_LENGTH)}
                 />
               </div>
-            </div>
-          ))}
-          {/* Show a blurb if there are more results than we are displaying */}
-          {results.length > resultsToShow && (
-            <div
-              className="animate-in fade-in slide-in-from-left-2 duration-150"
-              style={{
-                animationDelay: `${
-                  Math.min(resultsToShow, results.length) * 30
-                }ms`,
-                animationFillMode: "backwards",
-              }}
-            >
-              <div className="text-xs">
+            ))}
+            {/* Show a blurb if there are more queries than we are displaying */}
+            {queries.length > queriesToShow && (
+              <div
+                className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+                style={{
+                  animationDelay: `${queriesToShow * 30}ms`,
+                  animationFillMode: "backwards",
+                }}
+              >
                 <SourceChip2
-                  title={`${results.length - resultsToShow} more...`}
+                  title={`${queries.length - queriesToShow} more...`}
                   onClick={() => {
-                    setResultsToShow((prevResults) =>
+                    setQueriesToShow((prevQueries) =>
                       Math.min(
-                        prevResults + RESULTS_PER_EXPANSION,
-                        results.length
+                        prevQueries + QUERIES_PER_EXPANSION,
+                        queries.length
                       )
                     );
                   }}
                 />
               </div>
-            </div>
-          )}
+            )}
+          </div>
 
-          {/* If no results, and queries are showing, show a loading state */}
-          {results.length === 0 && queries.length > 0 && <BlinkingDot />}
+          {/* If no queries, show a loading state */}
+          {queries.length === 0 && <BlinkingDot />}
         </div>
+
+        {/* Only show results section for internal search, not web search */}
+        {!isInternetSearch && (
+          <div className="flex flex-col mt-3">
+            <div className="flex flex-wrap gap-x-2 gap-y-2 ml-1">
+              {results.slice(0, resultsToShow).map((result, index) => (
+                <div
+                  key={result.document_id}
+                  className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+                  style={{
+                    animationDelay: `${index * 30}ms`,
+                    animationFillMode: "backwards",
+                  }}
+                >
+                  <SourceChip2
+                    icon={<ResultIcon doc={result} size={10} />}
+                    title={truncateString(
+                      result.semantic_identifier || "",
+                      MAX_TITLE_LENGTH
+                    )}
+                    onClick={() => {
+                      if (result.link) {
+                        window.open(result.link, "_blank");
+                      }
+                    }}
+                  />
+                </div>
+              ))}
+              {/* Show a blurb if there are more results than we are displaying */}
+              {results.length > resultsToShow && (
+                <div
+                  className="text-xs animate-in fade-in slide-in-from-left-2 duration-150"
+                  style={{
+                    animationDelay: `${
+                      Math.min(resultsToShow, results.length) * 30
+                    }ms`,
+                    animationFillMode: "backwards",
+                  }}
+                >
+                  <SourceChip2
+                    title={`${results.length - resultsToShow} more...`}
+                    onClick={() => {
+                      setResultsToShow((prevResults) =>
+                        Math.min(
+                          prevResults + RESULTS_PER_EXPANSION,
+                          results.length
+                        )
+                      );
+                    }}
+                  />
+                </div>
+              )}
+
+              {/* If no results, and queries are showing, show a loading state */}
+              {results.length === 0 && queries.length > 0 && <BlinkingDot />}
+            </div>
+          </div>
+        )}
       </div>
     ),
   });
